Add a check to unfeasible analyzer : mainly extract class VariablesBoundsConsistency in separate source files
