+++ .github/workflows/windows-vcpkg.yml
@@ -89,7 +89,7 @@ jobs:
       uses: actions/setup-python@v4
       with:
         architecture: 'x64'
-        python-version: '3.11'
+        python-version: '3.12'
 
     - name: Install pip dependencies if necessary
       run: pip install -r src/tests/examples/requirements.txt
