+++ src/solver/infeasible-problem-analysis/unfeasible-pb-analyzer.h
@@ -10,32 +10,6 @@
 namespace Antares::Optimization
 {
 
-struct VariableBounds
-{
-    VariableBounds(std::string var_name, double low_bound, double up_bound) 
-        : name(var_name), lowBound(low_bound), upBound(up_bound)
-    {}
-
-    std::string name;
-    double lowBound;
-    double upBound;
-};
-
-class VariablesBoundsConsistency : public UnfeasibilityAnalysis
-{
-    using UnfeasibilityAnalysis::UnfeasibilityAnalysis;
-public:
-    void run() override;
-    void printReport() override;
-    std::string title() override { return "Variables bounds check"; }
-
-private:
-    void storeIncorrectVariable(std::string name, double lowBound, double upBound);
-    bool foundIncorrectVariables();
-
-    std::vector<VariableBounds> incorrectVars_;
-};
-
 class UnfeasiblePbAnalyzer
 {
 public:
