+++ src/solver/infeasible-problem-analysis/variables-bounds-consistency.cpp
@@ -1,0 +1,41 @@
+#include "variables-bounds-consistency.h"
+#include <antares/logs/logs.h>
+
+namespace Antares::Optimization
+{
+    
+void VariablesBoundsConsistency::run()
+{
+    for (auto& var : problem_->variables())
+    {
+        double lowBound = var->lb();
+        double upBound = var->ub();
+        std::string name = var->name();
+        if (lowBound > upBound)
+        {
+            storeIncorrectVariable(name, lowBound, upBound);
+        }
+    }
+
+    if (foundIncorrectVariables())
+        hasDetectedInfeasibilityCause_ = true;
+}
+
+void VariablesBoundsConsistency::storeIncorrectVariable(std::string name, double lowBound, double upBound)
+{
+    incorrectVars_.push_back(VariableBounds(name, lowBound, upBound));
+}
+
+bool VariablesBoundsConsistency::foundIncorrectVariables()
+{
+    return !incorrectVars_.empty();
+}
+
+void VariablesBoundsConsistency::printReport()
+{
+    for (auto& var : incorrectVars_)
+    {
+        logs.notice() << var.name << " : low bound = " << var.lowBound << ", up bound = " << var.upBound;
+    }
+}
+}   // namespace Antares::Optimization
